Split up remote/provider structs
